allow menu on linux too
