+++ pkg/client/tray_commands.go
@@ -1,5 +1,4 @@
-//go:build darwin || windows
-// +build darwin windows
+//go:build darwin || windows || linux
 
 package client
